python_api/requirements.txt
<<<<<<< Updated upstream
flask
tensorflow
pillow
numpy
h5py
=======
flask==3.0.3
tensorflow==2.17.1
keras==3.4.1
pillow==10.4.0
numpy==1.26.4
h5py==3.12.1
gunicorn==22.0.0
>>>>>>> Stashed changes
